requirements.txt
aiohttp
aiosignal
anyio
apify-client
apify-shared
appdirs
async-timeout
attrs
autopep8
beautifulsoup4
black
blinker
bs4
cachetools
certifi
charset-normalizer
click
docker-pycreds
et-xmlfile
exceptiongroup
Flask
frozenlist
gitdb
GitPython
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
googlesearch-python
h11
httpcore
httplib2
httpx
idna
itsdangerous
Jinja2
joblib
MarkupSafe
multidict
mypy-extensions
nltk
numpy
oauthlib
openai
openpyxl
packaging
pandas
pandas-stubs
pathspec
pathtools
platformdirs
protobuf
psutil
pyasn1
pyasn1-modules
pycodestyle
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
requests-oauthlib
rsa
sentry-sdk
setproctitle
six
smmap
sniffio
soupsieve
toml
tomli
tqdm
typing_extensions
tzdata
uritemplate
urllib3
wandb
Werkzeug
yarl
flask-sqlalchemy
flask-migrate
flask-cors
psycopg2-binary
alembic
